# api/openapi.yaml
openapi: 3.0.3
info:
  title: Idelium API Go
  version: 0.1.0
  description: >-
    Contract for the incremental Go replacement of the Idelium API. Business
    routes are added only after compatibility and tenant-isolation review.
servers:
  - url: /
paths:
...
  /admin/platforms/types:
    get:
      operationId: listPlatformTypes
      summary: List platform target types
      tags: [Platform Catalog]
      responses:
        "200":
          description: Legacy-compatible platform type array.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformCatalogItem"
        "500":
          description: The platform catalog is unavailable.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/platforms/status:
    get:
      operationId: listPlatformStatuses
      summary: List platform statuses
      tags: [Platform Catalog]
      responses:
        "200":
          description: Legacy-compatible platform status array.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformCatalogItem"
        "500":
          description: The platform catalog is unavailable.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ideliumcl/testcycle/{idTestCycle}:
    get:
      operationId: getCLITestCycle
      summary: Read one CLI test-cycle configuration
      description: >-
        Returns the Laravel-compatible test-cycle row for the authenticated
        customer. Missing and cross-tenant identifiers both return the legacy
        `Invalid id` response to avoid revealing foreign tenant resources.
      tags: [CLI Configuration]
      x-idelium-laravel-route: "/api/ideliumcl/testcycle/{idTestCycle}"
      x-idelium-controller: "App\\Http\\Controllers\\IdeliumClController@getTestCycle"
      x-idelium-current-owner: "go-application"
      x-idelium-trust-path: "api-key"
      x-idelium-authentication-mode: "api-key"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-cli", "idelium-docker-wiki"]
      parameters:
        - name: idTestCycle
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Laravel-compatible test-cycle row.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLITestCycle"
        "401":
          description: The legacy customer API key is missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "404":
          description: The test cycle is missing or outside the customer tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "500":
          description: The CLI configuration could not be loaded.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ideliumcl/test/{idTest}:
    get:
      operationId: getCLITest
      summary: Read one CLI test configuration
      description: >-
        Returns the Laravel-compatible test row for the authenticated customer.
        Missing and cross-tenant identifiers both return the legacy `Invalid id`
        response to avoid revealing foreign tenant resources.
      tags: [CLI Configuration]
      x-idelium-laravel-route: "/api/ideliumcl/test/{idTest}"
      x-idelium-controller: "App\\Http\\Controllers\\IdeliumClController@getTest"
      x-idelium-current-owner: "go-application"
      x-idelium-trust-path: "api-key"
      x-idelium-authentication-mode: "api-key"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-cli", "idelium-docker-wiki"]
      parameters:
        - name: idTest
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Laravel-compatible test row.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLITest"
        "401":
          description: The legacy customer API key is missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "404":
          description: The test is missing or outside the customer tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "500":
          description: The CLI configuration could not be loaded.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ideliumcl/step/{idStep}:
    get:
      operationId: getCLIStep
      summary: Read one CLI step configuration
      description: >-
        Returns the Laravel-compatible step row for the authenticated customer.
        Missing and cross-tenant identifiers both return the legacy `Invalid id`
        response to avoid revealing foreign tenant resources.
      tags: [CLI Configuration]
      x-idelium-laravel-route: "/api/ideliumcl/step/{idStep}"
      x-idelium-controller: "App\\Http\\Controllers\\IdeliumClController@getStep"
      x-idelium-current-owner: "go-application"
      x-idelium-trust-path: "api-key"
      x-idelium-authentication-mode: "api-key"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-cli", "idelium-docker-wiki"]
      parameters:
        - name: idStep
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Laravel-compatible step row.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLIStep"
        "401":
          description: The legacy customer API key is missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "404":
          description: The step is missing or outside the customer tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "500":
          description: The CLI configuration could not be loaded.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ideliumcl/plugins/{idProject}:
    get:
      operationId: listCLIPlugins
      summary: List CLI plugins for one project
      description: >-
        Returns the Laravel-compatible plugin array for the authenticated
        customer and requested project. Empty projects return an empty JSON
        array, preserving the legacy CLI contract.
      tags: [CLI Configuration]
      x-idelium-laravel-route: "/api/ideliumcl/plugins/{idProject}"
      x-idelium-controller: "App\\Http\\Controllers\\IdeliumClController@getPlugins"
      x-idelium-current-owner: "go-application"
      x-idelium-trust-path: "api-key"
      x-idelium-authentication-mode: "api-key"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-cli", "idelium-docker-wiki"]
      parameters:
        - name: idProject
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Laravel-compatible plugin array.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CLIPlugin"
        "401":
          description: The legacy customer API key is missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "404":
          description: The project identifier is malformed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "500":
          description: The CLI configuration could not be loaded.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ideliumcl/plugin/{idPlugin}:
    get:
      operationId: getCLIPlugin
      summary: Read one CLI plugin configuration
      description: >-
        Returns the Laravel-compatible plugin row for the authenticated
        customer. Missing and cross-tenant identifiers both return the legacy
        `Invalid id` response to avoid revealing foreign tenant resources.
      tags: [CLI Configuration]
      x-idelium-laravel-route: "/api/ideliumcl/plugin/{idPlugin}"
      x-idelium-controller: "App\\Http\\Controllers\\IdeliumClController@getPlugin"
      x-idelium-current-owner: "go-application"
      x-idelium-trust-path: "api-key"
      x-idelium-authentication-mode: "api-key"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-cli", "idelium-docker-wiki"]
      parameters:
        - name: idPlugin
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Laravel-compatible plugin row.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLIPlugin"
        "401":
          description: The legacy customer API key is missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "404":
          description: The plugin is missing or outside the customer tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyMessageResponse"
        "500":
          description: The CLI configuration could not be loaded.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  # BEGIN GENERATED LARAVEL COMPATIBILITY CONTRACTS
...
  /admin/platforms/brands:
    get:
      operationId: getAdminPlatformsBrands
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/brands"
      x-idelium-controller: "App\\Http\\Controllers\\BrandDeviceController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
    post:
      operationId: postAdminPlatformsBrands
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/brands"
      x-idelium-controller: "App\\Http\\Controllers\\BrandDeviceController@store"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
    put:
      operationId: putAdminPlatformsBrands
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/brands"
      x-idelium-controller: "App\\Http\\Controllers\\BrandDeviceController@update"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
...
  /admin/platforms/browsers/{idOs}:
    get:
      operationId: getAdminPlatformsBrowsersIdOs
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/browsers/{idOs}"
      x-idelium-controller: "App\\Http\\Controllers\\BrowserController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      parameters:
        - name: idOs
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
...
  /admin/platforms/browserversions/{idBrowser}:
    get:
      operationId: getAdminPlatformsBrowserversionsIdBrowser
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/browserversions/{idBrowser}"
      x-idelium-controller: "App\\Http\\Controllers\\VersionBrowserController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      parameters:
        - name: idBrowser
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
  /admin/platforms/locations:
    get:
      operationId: getAdminPlatformsLocations
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/locations"
      x-idelium-controller: "App\\Http\\Controllers\\LocationController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
    post:
      operationId: postAdminPlatformsLocations
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/locations"
      x-idelium-controller: "App\\Http\\Controllers\\LocationController@store"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
    put:
      operationId: putAdminPlatformsLocations
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/locations"
      x-idelium-controller: "App\\Http\\Controllers\\LocationController@update"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/LegacyCompatibilityPayload"
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
...
  /admin/platforms/models/{idBrand}:
    get:
      operationId: getAdminPlatformsModelsIdBrand
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/models/{idBrand}"
      x-idelium-controller: "App\\Http\\Controllers\\ModelDeviceController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      parameters:
        - name: idBrand
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
...
  /admin/platforms/os/{idType}:
    get:
      operationId: getAdminPlatformsOsIdType
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/os/{idType}"
      x-idelium-controller: "App\\Http\\Controllers\\OsController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      parameters:
        - name: idType
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
...
  /admin/platforms/osversion/{idOs}:
    get:
      operationId: getAdminPlatformsOsversionIdOs
      summary: Laravel compatibility contract
      description: >-
        Placeholder contract for an externally visible Laravel route. It keeps
        consumers, authentication, tenant scope, and migration ownership visible
        until this route is implemented natively in Go.
      tags: [Laravel Compatibility]
      x-idelium-laravel-route: "/api/admin/platforms/osversion/{idOs}"
      x-idelium-controller: "App\\Http\\Controllers\\VersionOsController@index"
      x-idelium-current-owner: "laravel-application"
      x-idelium-trust-path: "browser-session"
      x-idelium-authentication-mode: "browser-session"
      x-idelium-tenant-context: true
      x-idelium-consumers: ["idelium-web"]
      parameters:
        - name: idOs
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/LegacyCompatibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/LegacyCompatibilityError"
...
components:
  headers:
    CorrelationID:
      description: >-
        Validated request correlation identifier. A caller-provided identifier
        is preserved only when it contains 1-128 safe ASCII characters;
        otherwise the API generates a replacement.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
...
  responses:
    LegacyCompatibilityResponse:
      description: >-
        Laravel-compatible response whose exact body is captured by sanitized
        golden fixtures before native Go ownership.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LegacyCompatibilityPayload"
    LegacyCompatibilityError:
      description: Laravel-compatible error response.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is missing or invalid.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated principal cannot access this tenant resource.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: The request failed Laravel-compatible validation.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    LegacyCompatibilityPayload:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            correlationId:
              type: string
    LegacyMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    CLITestCycle:
      type: object
      required: [id, name, description, config, idProject, idCostumer]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        config:
          type: string
        idProject:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        idCostumer:
          type: integer
          format: int64
    CLITest:
      type: object
      required: [id, name, description, config, idProject, idCostumer]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        config:
          type: string
        idProject:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        idCostumer:
          type: integer
          format: int64
    CLIStep:
      type: object
      required: [id, name, description, config, idProject, order, idCostumer]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        config:
          type: string
        idProject:
          type: integer
          format: int64
        order:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        idCostumer:
          type: integer
          format: int64
    CLIPlugin:
      type: object
      required: [id, name, code, description, idProject, idCostumer]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        description:
          type: string
        idProject:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        idCostumer:
          type: integer
          format: int64
    PlatformCatalogItem:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string